Exclusão de cliente, lançando exceção de negócio.

+++ src/main/java/model/bo/ClienteBO.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 
 import model.dao.ClienteDAO;
 import model.entity.Cliente;
+import model.exception.ClienteComLinhaTelefonicaException;
 
 public class ClienteBO {
 
@@ -32,9 +33,32 @@ public class ClienteBO {
 		return dao.consultarTodos();
 	}
 
-	public boolean excluir(Cliente clienteParaExcluir) {
-		// TODO validar se o cliente tem linha telefônica
+	/**
+	 * Exclui um cliente, somente se ele não possuir nenhuma linha telefônica
+	 * @param clienteParaExcluir o cliente a ser excluído
+	 * @return se excluiu ou não
+	 * @throws ClienteComLinhaTelefonicaException lançada quando o cliente possui linha(s)
+	 */
+	public boolean excluir(Cliente clienteParaExcluir) throws ClienteComLinhaTelefonicaException {
+		boolean excluiu;
 		
-		return dao.remover(clienteParaExcluir.getId());
+		if(clienteParaExcluir.getLinhas().size() > 0) {
+			throw new ClienteComLinhaTelefonicaException("Cliente informado não pode ser excluído,"
+					+ " pois possui linha(s) telefônica(s)");
+		}else {
+			excluiu = dao.remover(clienteParaExcluir.getId());
+		}
+		
+		return excluiu;
 	}
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 }
